Moved research tm_cpp_test.py back into nupic.research
